Simplify pdf_utils.py by removing class, just functions
