+++ fuzzing/extract_text_to_fp_fuzzer.py
@@ -3,10 +3,15 @@ import sys
 
 import atheris
 
-from fuzzing.fuzz_helpers import EnhancedFuzzedDataProvider
+from fuzzing.fuzzed_data_provider import PdfminerFuzzedDataProvider
 
 with atheris.instrument_imports():
-    from .pdf_utils import PDFValidator, prepare_pdfminer_fuzzing
+    from fuzzing.utils import (
+        prepare_pdfminer_fuzzing,
+        is_valid_byte_stream,
+        generate_layout_parameters,
+        should_ignore_error,
+    )
     from pdfminer.high_level import extract_text_to_fp
     from pdfminer.psparser import PSException
 
@@ -15,11 +20,11 @@ available_layout_modes = ["exact", "normal", "loose"]
 
 
 def fuzz_one_input(data: bytes) -> None:
-    if not PDFValidator.is_valid_byte_stream(data):
+    if not is_valid_byte_stream(data):
         # Not worth continuing with this test case
         return
 
-    fdp = EnhancedFuzzedDataProvider(data)
+    fdp = PdfminerFuzzedDataProvider(data)
 
     try:
         with fdp.ConsumeMemoryFile(all_data=False) as f_in, io.BytesIO() as f_out:
@@ -27,7 +32,7 @@ def fuzz_one_input(data: bytes) -> None:
                 f_in,
                 f_out,
                 output_type=fdp.PickValueInList(available_output_formats),
-                laparams=PDFValidator.generate_layout_parameters(fdp),
+                laparams=generate_layout_parameters(fdp),
                 maxpages=fdp.ConsumeIntInRange(0, 10),
                 page_numbers=fdp.ConsumeOptionalIntList(10, 0, 10),
                 scale=fdp.ConsumeFloatInRange(0.0, 2.0),
@@ -38,7 +43,7 @@ def fuzz_one_input(data: bytes) -> None:
     except (AssertionError, PSException):
         return
     except Exception as e:
-        if PDFValidator.should_ignore_error(e):
+        if should_ignore_error(e):
             return
         raise e
